gets the ekman case workin
